Speedup outdated parts loading (#49317)
